mcp-server/ingestion/requirements.txt
sentence-transformers==2.3.1
PyPDF2==3.0.1
pdfplumber==0.10.3
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
tqdm==4.66.1
